Создан модуль для компонента главного меню
